+++ __test__/__mockdata__/auth/index.js
@@ -16,6 +16,14 @@ export const data = {
       password: 'thisismypassword',
     },
 
+    inactiveSignup: {
+      first_name: 'Macheza',
+      last_name: 'Dzabala',
+      phonenumber: '0991807269',
+      email: 'demouser@gmail.com',
+      password: 'thisismypassword',
+    },
+
     missingEmail: {
       first_name: 'Macheza',
       last_name: 'Dzabala',
