+++ __test__/index.spec.js
@@ -2,7 +2,8 @@ import signUpTests from './__auth__/signupTest.spec';
 import signInTests from './__auth__/signinTest.spec';
 import getOnePostTests from './__posts__/getOnePostTest.spec';
 import postTests from './__posts__/postTests';
-import adminTest from './__admin__/adminTest.spec';
+import categoryTests from './__admin__/categoryTest.spec';
+import adminTests from './__admin__/adminTests.spec';
 import deletePost from './__posts__/deletePost.spec';
 
 describe('Test Runner', () => {
@@ -10,7 +11,8 @@ describe('Test Runner', () => {
   describe('Signin Tests', signInTests);
   describe('Get One Post Tests', getOnePostTests);
   describe('Create New Post Test', postTests);
-  describe('Create New category', adminTest);
+  describe('Category Tests', categoryTests);
+  describe('Admin Tests', adminTests);
   describe('Post tests', postTests);
   describe('Delete Post', deletePost);
 });
